+++ src/js/ui/react/index.js
@@ -9,11 +9,17 @@
  * @author Mikel Tuesta <mikeltuesta@gmail.com>
  */
 
+import ArrowDown from './svg/ArrowDown';
+import ResizeAware from './hoc/ResizeAware';
+import Loader from './atoms/Loader/Loader';
 import FormSelect from './atoms/FormSelect/FormSelect';
 import FormGroupInput from './components/FormGroupInput/FormGroupInput';
 import FormGroupSelect from './components/FormGroupSelect/FormGroupSelect';
 
 export {
+  ArrowDown,
+  ResizeAware,
+  Loader,
   FormSelect,
   FormGroupInput,
   FormGroupSelect